requirements.txt
diffusers~=0.32.2
torch~=2.6.0
torchaudio
torchvision
huggingface-hub~=0.29.3
python-dotenv~=1.0.0